pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["yang"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.4"
pyqtgraph = "0.13.3"
PyQt5 = "5.15.10"
pyqt5-qt5 = "5.15.2"
pyserial = "^3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"
